requirements.txt
telethon>=1.34.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
convex>=0.7.0
